whatsapp-message/requirements.txt
playwright>=1.40
anthropic>=0.40
twilio>=9.0
python-dotenv>=1.0
beautifulsoup4>=4.12
python-dateutil>=2.8
pytest>=8.0
pytest-asyncio>=0.23
